Change configuration to deal w macOS
On macOS it seems that /var/run/hugo gets deleted upon a reboot, so my
approach of creating it once won't work.  This is probably too much
hassle to figure out how to fix a different way, so I made the default
configuration use a local run directory in the root of the local
directory.  Also, this means the files checked into git shouldn't
default to having the development server address in it, so I removed
that as well.