Fix scaling of icon relative to page
There's some weirdness in my css that causes step changes in page
width as the font size is changed, and using something like 10% of the
width means that the icon would sometimes get _smaller_ when you made
the font bigger.  Therefore, instead of scaling as percentage of page
width, we should be scaling relative to the font size.